% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.KFold}
\alias{CV.KFold}
\title{Cross-Validation with K-Folds}
\usage{
CV.KFold(DataSet, K = 5, Set_seed = NULL)
}
\arguments{
\item{DataSet}{\code{data.frame} Tidy data.frame, with $Line especified on it, also works taking account $Env and $Trait.}

\item{K}{\code{integer} Number of folds.}

\item{Set_seed}{\code{integer} Number of seed}
}
\value{
\code{List} A list object with lenght of
}
\examples{
\dontrun{
  library(IBCF.MTME)
  data('Wheat_IBCF')
  DataSet <- data.frame(Data.Trigo_IBCF[,-1])
  DataSet <- Matrix2Tidy(DataSet)
  CV.KFold(DataSet = DataSet, K = 5, Set_seed = 10)
}
}
