% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBCFData.R
\name{IBCF.Data}
\alias{IBCF.Data}
\title{Prepare and validate the data for the IBCF Model}
\usage{
IBCF.Data(ID = "", Environment = "", Trait = "", Response = "",
  data = NULL, messages = TRUE, get = "all")
}
\arguments{
\item{ID}{The identifier of the observation}

\item{Environment}{The name of the environments where the observation was evaluated}

\item{Trait}{The name of the traits that was measured in the environment.}

\item{Response}{The measure of the trait.}

\item{data}{(data.frame object) The dataset where can be recovered the data.}

\item{messages}{(Logical) Should be printed the messages of the function?}

\item{get}{(string) By default the function return 'all' the transformations of the data, if you want only the tidy format of the data select 'tidy' or 'matrix' if you want only the matrix format.}
}
\value{
one or two list IBCFData object.
}
\examples{
\dontrun{
  library(IBCF.MTME)

  ### Choose one
  data <- data.frame(Response = c(1,23,4,25,23,5))
  data <- data.frame(Response = c(1,23,4,25,23,5), Environment = c('E1', 'E1', 'E2', 'E1', 'E2', 'E2'))
  data <- data.frame(Response = c(1,23,4,25,23,5), Trait = c('T1', 'T1', 'T2', 'T1', 'T2', 'T2_'))
  data <- data.frame(Response = c(1,23,4,25,23,5,6,79), Trait = c('T1', 'T1', 'T2', 'T2', 'T2', 'T2', 'T1', 'T1_'), Environment = c('E1_', 'E2', 'E2', 'E1', 'E2', 'E1', 'E2', 'E1'))
  data <- data.frame(Response = c(1,23,4,25,23,5,6), Trait = c('T1', 'T1', 'T2', 'T2', 'T2', 'T2', 'T1'), Environment = c('E1_', 'E2', 'E2', 'E1', 'E2', 'E1', 'E2'))

  ## Validate the dataset and obtain only the matrix format.
  Data <- IBCF.Data(Environment = 'Environment', Trait = 'Trait', Response = 'Response', data = data, get = 'Matrix')  # Could change get to 'all' or 'tidy'
}

}
