% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBCFYear.R
\name{IBCF.Years}
\alias{IBCF.Years}
\title{IBCF.Years}
\usage{
IBCF.Years(DataSet, Years.testing = "", Traits.testing = "", dec = 4)
}
\arguments{
\item{DataSet}{\code{data.frame}}

\item{Years.testing}{\code{vector}}

\item{Traits.testing}{\code{vector}}
}
\value{
A list with the next components
\item{Year.testing}{\code{vector} a vector with the Years used for the testing data}
\item{Traits.testing}{\code{vector} a vector with the Traits used for the testing data}
\item{Data_Obs_Pred}{\code{data.frame} Contains the values observed and predicted (the predicted values has '.1' after the name)}
\item{predictions_Summary}{\code{data.frame} Contains the summary of the correlation of the predictions and de mean squared error of the predictions (MSEP)}
}
\description{
Item Based Collaborative Filterign for Years data
}
\examples{


}
