% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.IBCF}
\alias{plot.IBCF}
\title{Plot IBCF graph}
\usage{
\method{plot}{IBCF}(x, select = "Pearson", ...)
}
\arguments{
\item{object}{\code{IBCF object} Objeto IBCF, resultado de ejecutar IBCF()}
}
\description{
Graphs from IBCF object
}
