% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.IBCFY}
\alias{plot.IBCFY}
\title{plot.IBCFY}
\usage{
\method{plot}{IBCFY}(x, select = "Pearson", ...)
}
\arguments{
\item{object}{\code{IBCFY object} Objeto IBCFY, resultado de ejecutar IBCF.Year()}
}
\description{
Solo es una prueba
}
