% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.IBCF}
\alias{print.IBCF}
\title{Print IBCF information object}
\usage{
\method{print}{IBCF}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print IBCF information object
}
