% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.IBCFY}
\alias{print.IBCFY}
\title{Print IBCFY information object}
\usage{
\method{print}{IBCFY}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print IBCFY information object
}
