% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.IBCF}
\alias{residuals.IBCF}
\title{residuals.IBCF}
\usage{
\method{residuals}{IBCF}(object, ...)
}
\arguments{
\item{object}{\code{IBCF object} Objeto IBCF, resultado de ejecutar IBCF}
}
\description{
Solo es una prueba
}
