% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.IBCF}
\alias{summary.IBCF}
\title{Summary.IBCF}
\usage{
\method{summary}{IBCF}(object, ...)
}
\arguments{
\item{object}{\code{IBCF object} Objeto IBCF, resultado de ejecutar IBCF}
}
\description{
Solo es una prueba
}
