% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.IBCFY}
\alias{summary.IBCFY}
\title{Summary}
\usage{
\method{summary}{IBCFY}(object, ...)
}
\arguments{
\item{object}{\code{IBCFY object} Objeto IBCFY, resultado de ejecutar IBCF.Years}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of IBCFY object
}
