% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Td2M.R
\name{Td2M}
\alias{Td2M}
\alias{Td2M}
\title{TidyData to Matrix}
\usage{
Td2M(DataSet)

Td2M(DataSet)
}
\arguments{
\item{DataSet}{Dataset}

\item{DataSet}{Dataset}
}
\description{
TidyData to Matrix

Matrix to TidyData
}
