% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_IBCF.R
\docType{data}
\name{Wheat_IBCF}
\alias{Wheat_IBCF}
\title{Wheat Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\source{
Not yet
}
\usage{
data(Wheat_IBCF)
}
\description{
This data needs a description
}
\author{
Not yet
}
