% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year_IBCF.R
\docType{data}
\name{Year_IBCF}
\alias{Year_IBCF}
\title{Year_IBCF Data}
\format{a \code{TidyData} format, 750 row per 4 columns.}
\source{
Not yet
}
\usage{
data(Year_IBCF)
}
\description{
This data needs a description
}
\author{
Not yet
}
