% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{barplot.IBCFY}
\alias{barplot.IBCFY}
\title{barplot.IBCFY}
\usage{
\method{barplot}{IBCFY}(height, select = "Pearson", ...)
}
\arguments{
\item{height}{\code{IBCFY object} IBCFY object, result of use the IBCF.Years() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the IBCF Object, else ('MAAPE'), plot the MAAPE of the IBCF Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Barplot of the results from IBCFY object
}
