% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.RandomPart}
\alias{CV.RandomPart}
\title{Cross-Validation with Random Partions}
\usage{
CV.RandomPart(DataSet, NPartitions = 10, PTesting = 0.35, Set_seed = NULL)
}
\arguments{
\item{DataSet}{\code{data.frame} Number of Lines in the DataSet. By default are 10 Lines.}

\item{NPartitions}{\code{integer} Number of Partitions for the Cross-Validation. Is 10 by default.}

\item{PTesting}{\code{Double} Porcentage of Testing for the Cross-Validation. Is 0.35 by default.}

\item{Set_seed}{\code{integer} Number of seed for replicable research. Is NULL by default.}
}
\value{
\code{List} A list object with lenght of \code{NPartitions}, every indice has a \code{matrix} \eqn{n \times x}, where \eqn{n} is the number of \code{NLines} and \eqn{x} is the number of  \code{NEnv} \eqn{\times} \code{NTraits}. The values inside is 1 for training and 2 for testing.
}
\examples{
\dontrun{
  CV.RandomPart(DataSet)
  CV.RandomPart(DataSet, NPartitions = 10)
  CV.RandomPart(DataSet, NPartitions = 10, PTesting = .35)
  CV.RandomPart(DataSet, NPartitions = 10, PTesting = .35, Set_seed = 5)
}
}
