% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBCF.R
\name{IBCF}
\alias{IBCF}
\title{IBCF: Item Based Collaborative Filterign}
\usage{
IBCF(DataSet, CrossValidation = NULL)
}
\arguments{
\item{DataSet}{\code{data.frame} Sets of data...}

\item{CrossValidation}{\code{list} List with the partitions}
}
\description{
Item Based Collaborative Filterign for multi-trait and multi-environment data.
}
\examples{

}
