% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.IBCF}
\alias{plot.IBCF}
\title{Plot IBCF graph}
\usage{
\method{plot}{IBCF}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{IBCF object} Objeto IBCF, resultado de ejecutar IBCF()}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the IBCF Object, else ('MSEP'), plot the MSEP of the IBCF Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Graphs from IBCF object
}
