% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.IBCF}
\alias{plot.IBCF}
\title{Plot IBCF graph}
\usage{
\method{plot}{IBCF}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{IBCF object} IBCF object, result of use the IBCF() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the IBCF Object, else ('MAAPE'), plot the MAAPE of the IBCF Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from IBCF object
}
