% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.IBCF}
\alias{print.IBCF}
\title{Print IBCF information object}
\usage{
\method{print}{IBCF}(x, ...)
}
\arguments{
\item{x}{IBCF object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
printeable object
}
\description{
Print IBCF information object
}
