% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.IBCF}
\alias{summary.IBCF}
\title{Summary}
\usage{
\method{summary}{IBCF}(object, ...)
}
\arguments{
\item{object}{\code{IBCF object} Objeto IBCF, resultado de ejecutar IBCF}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of IBCF object
}
