% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.IBCF}
\alias{summary.IBCF}
\title{Summary}
\usage{
\method{summary}{IBCF}(object, information = "compact", digits = 4,
  ...)
}
\arguments{
\item{object}{\code{IBCF object} IBCF object, result of use the IBCF() function}

\item{information}{\code{string} ...}

\item{digits}{\code{numeric} ...}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of IBCF object
}
