% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.IBCFY}
\alias{summary.IBCFY}
\title{Summary}
\usage{
\method{summary}{IBCFY}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{IBCFY object} IBCFY object, result of use the IBCF.Years() function}

\item{digits}{\code{numeric} Number of digits of the output.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of IBCFY object
}
