% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_IBCF.R
\docType{data}
\name{Wheat_IBCF}
\alias{Wheat_IBCF}
\title{Wheat Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\usage{
data(Wheat_IBCF)
}
\description{
The package includes a data set based on a portion of the data used in the study of
Montesinos-Lopez, O. A.; Montesinos-Lopez, A.; Crossa, J.; Toledo, F. H.; Montesinos-Lopez, J. C.; Singh, P. & Salinas-Ruiz, J. (2017). A Bayesian Poisson-lognormal Model for Count Data for Multiple-Trait Multiple-Environment Genomic-Enabled Prediction. G3: Genes|Genomes|Genetics 7(5):1595–1606. http://doi.org/10.1534/g3.117.039974.
The data set consists of 250 wheat lines evaluated in 3 environments and 4 distinct traits, i.e. 3000 observations.
}
\author{
Montesinos-Lopez, O. A.
}
